Да

+++ CoffeeSh0p/Coffee.cs
@@ -4,11 +4,11 @@
     {
         public int quantity;
         public string coffeeName;
-        public short price;
+        public int price;
         public string sugar;
         public string cinnamon;
         public string sirop;
-        public Coffee(int quantity, string coffeeName, short price, string sugar, string cinnamon, string sirop)
+        public Coffee(int quantity, string coffeeName, int price, string sugar, string cinnamon, string sirop)
         {
             this.quantity = quantity;
             this.coffeeName = coffeeName;
@@ -17,7 +17,5 @@
             this.cinnamon = cinnamon;
             this.sirop = sirop;
         }
-        
-            
     }
 }
